Rename: ArtistList 경로 변경
