+++ client/src/index.tsx
@@ -16,7 +16,7 @@ import Artistregist from './pages/artist-register/ArtistRegister';
 import Cancel from './pages/cancel/Cancel';
 import SignUpPage from './pages/sign/SignUp';
 import SignInPage from './pages/sign/SignIn';
-import ArtistList from './pages/ArtistList';
+import ArtistList from './pages/artist-list/ArtistList';
 import Home from './pages/home/Home';
 import ReviewRegister from './pages/review/ReviewRegister';
